vdpa/ifc: fix reconnection in SW-assisted live migration
[ upstream commit ef3be7e237dd707f46870b09b93f88f8ee5e75c6 ]

In the case using argument "sw-live-migration=1" to enable SW assisted live
migration, we take QEMU as front end for example, after source VM migrates
to destination VM, we keep vdpa process for source VM there, we kill the
QEMU process for source VM, and restart the QEMU process for source VM.
In this case, vdpa driver will not perform DMA map and data path will not
work properly.

The above case works fine in the case "sw-live-migration=0".

The root cause is that current code driver does not set running flag to 0.
Driver treats device as ruuning and does not perform DMA map.

Fixes: 4bb531e152d3 ("net/ifc: support SW assisted VDPA live migration")

Signed-off-by: Andy Pei <andy.pei@intel.com>
Reviewed-by: Maxime Coquelin <maxime.coquelin@redhat.com>