cmdline: handle EOF as quit
[ upstream commit 415549f1ccce62b82cb182175346904a65f74cec ]

If end of file is reached on input, then cmdline_poll() will
return 1 (ie file has something); and then the cmdline_in()
call to read will return 0. With the existing code,
caller has no way to tell that end of file has been reached
 and will retry forever.

A good way to handle this is to make end of file equivalent
to the quit command. Since no more input is possible at that
point.

Fixes: 067855e651d6 ("cmdline: add polling mode")

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>