kvargs: add API documentation for process callback
[ upstream commit 52ab17efdecf935792ee1d0cb749c0dbd536c083 ]

The rte_kvargs_process() is used to parse KV pairs, it also supports
to parse 'only keys' (e.g. socket_id) type. And the callback function
(which prototype is arg_handler_t) parameter 'value' is NULL when
parsing 'only keys'.

But there is no detailed definition of 'value' may be NULL, so this
patch adds it.

Signed-off-by: Chengwen Feng <fengchengwen@huawei.com>
Acked-by: Olivier Matz <olivier.matz@6wind.com>