app/testpmd: revert primary process polling all queues fix
[ upstream commit d7d802daf80f4a03c77a815da6d47a7b3657ef2c ]

For some drivers [1], testpmd forwarding is broken with commit [2].

This is because with [2] testpmd gets queue state from ethdev and
forwarding is done only on queues in started state, but some drivers
don't update queue status properly, and this breaks forwarding for those
drivers.

Drivers should be fixed but more time is required to verify drivers
again, instead reverting [2] for now to not break drivers.
Target is to merge [2] back at the beginning of next release cycle and
fix drivers accordingly.

[1]
Bugzilla ID: 1259

[2]
Fixes: 141a520b35f7 ("app/testpmd: fix primary process not polling all queues")

Signed-off-by: Ferruh Yigit <ferruh.yigit@amd.com>
Tested-by: Song Jiale <songx.jiale@intel.com>
Tested-by: Ali Alnubani <alialnu@nvidia.com>