test: fix named test macro
[ upstream commit 63bf81a617ac25017e23441521743af7877a38fa ]

Currently, the TEST_CASE_NAMED_ST macro is implemented in a way that
is identical to TEST_CASE_NAMED_WITH_DATA. In particular, it expects
the test function to have a data argument, which is why a bunch of tests
have an unused data arguments.

Fix this by adjusting the macro to not expect a test function with data
arguments, and adjust the dependent test cases accordingly.

Fixes: 474f12024dd9 ("test/crypto: add lookaside IPsec cases")

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Acked-by: Akhil Goyal <gakhil@marvell.com>
Acked-by: Hemant Agrawal <hemant.agrawal@nxp.com>
Acked-by: Anoob Joseph <anoobj@marvell.com>