enetfec: Replace dcbf(dc ivac) with dc civac
Flush and invalidate the packet instead of just flushing so that next
time when same buffer is reused, application doesn't read the stale
cache entry.

This change will be applicable for only forwarded packet and not for
termination packet since termination packet won't be freed by driver
hence wont be flushed and invalidated.

Signed-off-by: Rohit Raj <rohit.raj@nxp.com>
DPDK-4300
DPDK-4307